backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
httpx==0.25.2
pydantic==2.5.2
pydantic-settings==2.1.0

# Database
supabase==2.0.3

# AI APIs
anthropic==0.7.8
google-generativeai==0.3.1
replicate==0.22.0

# Payments
stripe==7.8.0

# Storage
boto3==1.33.6  # S3-compatible for Cloudflare R2

# Utilities
python-multipart==0.0.6
aiofiles==23.2.1
Pillow==10.1.0  # Image processing for texture packs
zipfile36==0.1.3

# Background tasks
celery==5.3.4
redis==5.0.1
